+++ platform-operator/controllers/verrazzano/component/registry/registry_test.go
@@ -4,7 +4,8 @@
 package registry
 
 import (
-	"github.com/verrazzano/verrazzano-modules/pkg/controller/base/controllerspi"
+	"github.com/verrazzano/verrazzano-modules/pkg/controller/spi/controllerspi"
+	apiextensionsv1 "k8s.io/apiextensions-apiserver/pkg/apis/apiextensions/v1"
 	"reflect"
 	"testing"
 
@@ -891,6 +892,11 @@ func (f fakeComponent) ShouldUseModule() bool {
 // GetWatchDescriptors returns the list of WatchDescriptors for objects being watched by the component
 func (f fakeComponent) GetWatchDescriptors() []controllerspi.WatchDescriptor {
 	return nil
+}
+
+// GetModuleConfigAsHelmValues returns an unstructured JSON snippet representing the portion of the Verrazzano CR that corresponds to the module
+func (f fakeComponent) GetModuleConfigAsHelmValues(effectiveCR *v1alpha1.Verrazzano) (*apiextensionsv1.JSON, error) {
+	return nil, nil
 }
 
 func (f fakeComponent) GetJSONName() string {